+++ test/test_parameters.py
@@ -35,7 +35,7 @@ def test_path(testdir):
     # When parameter is specified for "path"
     testdir.make_test(
         """
-@schema.parametrize(filter_endpoint="/users/{user_id}")
+@schema.parametrize(endpoint="/users/{user_id}")
 @settings(max_examples=3)
 def test_(case):
     assert_int(case.path_parameters["user_id"])